Move FlyDSL decode kernels into decode/ subpackage and drop internal-reference comments
